% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04realization.R
\name{adj.true.CV}
\alias{adj.true.CV}
\title{True overall CV with bias adjustment}
\usage{
adj.true.CV(n.dat)
}
\arguments{
\item{n.dat}{new survey data sets at wave levels}
}
\value{
dataframe from wave 0 to the end
}
\description{
Need mircodata.
}
