% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02distribution.R
\name{pposterior}
\alias{pposterior}
\title{Beta posterior distributions}
\usage{
pposterior(n.dat, h.dat, s.score, a = 1, b = 1, svy.ref)
}
\arguments{
\item{n.dat}{new survey data sets at wave levels}

\item{h.dat}{historical survey data sets}

\item{s.score}{historcial-level similarity scores dependent on feature-level weights}

\item{a}{shape1 parameter with default value 1}

\item{b}{shape2 parameter with default value 1}

\item{svy.ref}{historical reference exists with 1 and 0 otherwise}
}
\value{
data frame with shape 1, shape 2, strata and wave
}
\description{
\code{pprior} parameters are updated to \code{pposterior} parameter wave by wave
and posterior will serve as the prior for next data collection wave.
}
