% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02distribution.R
\name{pprior}
\alias{pprior}
\title{Power prior with Beta form}
\usage{
pprior(n.hat, h.dat, s.score, a = 1, b = 1, svy.ref)
}
\arguments{
\item{n.hat}{new survey data sets at wave levels}

\item{h.dat}{historical survey data sets}

\item{s.score}{historcial-level similarity scores dependent on feature-level weights}

\item{a}{shape1 parameter with default value 1}

\item{b}{shape2 parameter with default value 1}

\item{svy.ref}{historical reference exists with 1 and 0 otherwise}
}
\value{
dataframe includes shape parameters of power prior
}
\description{
Given historical survey data sets, their posterior distribution
serves as the prior distribution to a new survey. The prior will be
updated to posterior when new data comes in.
}
