% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01split.R
\name{redat}
\alias{redat}
\title{Reshape survey data sets with respondents and sample size separately.}
\usage{
redat(svy.dat)
}
\arguments{
\item{svy.dat}{survey data sets}
}
\value{
a list
}
\description{
For new survey data sets are indicayed by wave, the iteration of new sample released.
For historical survey data sets are indicated by the historical names.
}
