% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03sampling.R
\name{rho.sample}
\alias{rho.sample}
\title{Empirical samples from response propensity}
\usage{
rho.sample(n.dat, h.dat, s.score, a = 1, b = 1, svy.ref, N = 10000)
}
\arguments{
\item{n.dat}{new survey data sets at wave levels}

\item{h.dat}{historical survey data sets}

\item{s.score}{historcial-level similarity scores dependent on feature-level weights}

\item{a}{shape1 parameter with default value 1}

\item{b}{shape2 parameter with default value 1}

\item{svy.ref}{historical reference exists with 1 and 0 otherwise}

\item{N}{the number of observations (replication) with default 10000}
}
\value{
a list from wave 0 to the end
}
\description{
Draw samples from power prior and updated posterior and later used to
evaluate quality indicators.
}
