% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04realization.R
\name{rho.true.w}
\alias{rho.true.w}
\title{True response propensities}
\usage{
rho.true.w(n.dat, h.dat, s.score, a = 1, b = 1, svy.ref)
}
\arguments{
\item{n.dat}{new survey data sets at wave levels}

\item{h.dat}{historical survey data sets}

\item{s.score}{historcial-level similarity scores dependent on feature-level weights}

\item{a}{shape1 parameter with default value 1}

\item{b}{shape2 parameter with default value 1}

\item{svy.ref}{historical reference exists with 1 and 0 otherwise}
}
\value{
data frame with true response propensity from wave 0 to data collection end
}
\description{
Based on observation realized stratum response propensities are derived
and used to calculate true overall response propensity.
}
