% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05estimated.R
\name{simulated.CV}
\alias{simulated.CV}
\title{Estimated CV}
\usage{
simulated.CV(n.dat, h.dat, s.score, s.w, a = 1, b = 1, svy.ref, N = 10000)
}
\arguments{
\item{n.dat}{new survey data sets at wave levels}

\item{h.dat}{historical survey data sets}

\item{s.score}{historcial-level similarity scores dependent on feature-level weights}

\item{s.w}{subgroup proportion}

\item{a}{shape1 parameter with default value 1}

\item{b}{shape2 parameter with default value 1}

\item{svy.ref}{historical reference exists with 1 and 0 otherwise}

\item{N}{the number of observations (replication) with default 10000}
}
\value{
dataframe with N replicated draws with relative to one wave
}
\description{
Estimated CV
}
