% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06performance.R
\name{sumstats}
\alias{sumstats}
\title{Prediction performance}
\usage{
sumstats(
  adjust = "no",
  simulated.fun,
  true.fun,
  n.dat,
  h.dat,
  s.score,
  s.w,
  a = 1,
  b = 1,
  svy.ref,
  N = 10000
)
}
\arguments{
\item{adjust}{CV with bias adjusted is 1 and 0 otherwise}

\item{simulated.fun}{simulated indicators}

\item{true.fun}{true indicators}

\item{n.dat}{new survey data sets at wave levels}

\item{h.dat}{historical survey data sets}

\item{s.score}{s.score historcial-level similarity scores dependent on feature-level weights}

\item{s.w}{subgroup proportion}

\item{a}{shape1 parameter with default value 1}

\item{b}{shape2 parameter with default value 1}

\item{svy.ref}{historical reference exists with 1 and 0 otherwise}

\item{N}{the number of observations (replication) with default 10000}
}
\value{
dataframe with posterior mean, variance, quantiles and resulting RMSE with regards to waves
}
\description{
Some statistics, including posterior mean, lower(upper) quantile and RMSE.
}
