% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04realization.R
\name{true.RR}
\alias{true.RR}
\title{True overall RR}
\usage{
true.RR(n.dat, h.dat, s.score, s.w, a = 1, b = 1, svy.ref)
}
\arguments{
\item{n.dat}{n.dat new survey data sets at wave levels}

\item{h.dat}{historical survey data sets}

\item{s.score}{historcial-level similarity scores dependent on feature-level weights}

\item{s.w}{subgroup proportion}

\item{a}{shape1 parameter with default value 1}

\item{b}{shape2 parameter with default value 1}

\item{svy.ref}{historical reference exists with 1 and 0 otherwise}
}
\value{
dataframe with wave-level true rr
}
\description{
The weighted response rates as first quality indicator.
}
